v1.15.0 -- DAQ changes: VFAT live words count, input autokill; TTC command buffer min max are 2 and 4 by default
DAQ input processor is now also counting the number of zero suppressed
VFAT words and putting that into the datastream, using a spot previously
dedicated to per chamber zero suppression flags (bits [51:40] in chamber
header). DAQ format version has been changed from 0 to 1 to reflect
that.
Also registers are available to read the current, min, and max "VFAT
live word count", which is the number of non-zero-suppressed +
zero-suppressed 64bit words (total should be equal to 24 * 3 in normal
conditions regardless of zero suppression setting)
Also this version introduces autokilling of bad links (is enabled by
default, but can be disabled). When a given link times out more than a
certain number of times in a row (this parameter is configurable, and is
set to 100 by default), then this input will be autokilled.
The autokill mask is reset during each resync. The autokill mask, and
the individual link timeout counters (both consecutive and total) can be
read with registers.
Default values of the TTC command buffer min and max for OOS have been
set to 2 and 4 respectively (previously these were set at 3 and 3).